re-introduce cutting guides now that front faces are all on one page
A first test showed that I'm losing up to 5mm of card space,
considerably altering the appearance of the cards...
Not sure this is the way to go. Might help to simply be more careful feeding the paper into the printer.
At least I assume it would work better if I gave this to a print-shop.