build: adjust project-specific limits and rules for pylint config
This commit is meant to be re-applied after pylint version is changed and the config re-generated.
Probably best workflow is to re-generate with old version, then generate with the new one to see which values are new and which defaults changed. Commit that, then cherry-pick this commit, resolve conflicts, adjust new values and commit that before addressing any new violations.